Fix: Enhance cscope search types and add transient menu
This commit enhances the `cscope-search-types` alist to include more
search options and associated keys.  It also introduces a transient menu
for easier access to cscope actions.

The `cscope-search-types` alist now includes the search key and
description. The `cscope-entry-actions` variable is introduced to define
the actions available in the transient menu. Keybindings are added to
`cscope-mode-map` for frequently used commands like database
regeneration, previous query, and next query.  A transient prefix
command `cscope-entry` is defined to present the cscope actions in a
menu.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>