Use consistent keybinding for cscope-generate-database
The transient menu for cscope had an inconsistent keybinding for
`cscope-generate-database`.  The keybinding in the menu was `g`
(lowercase), while the keybinding in the `cscope-mode-map` was `G`
(uppercase).

This commit changes the keybinding in the transient menu to `G` to match
the keybinding in the `cscope-mode-map`, providing a more consistent and
intuitive user experience.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>