Initialize cscope on load and improve initialization
This commit refactors the cscope initialization process to ensure that
interactive functions and transient menu functions are generated when
the module is loaded.

Previously, these functions were only generated when `cscope-mode` was
activated. This meant that users could not use the `cscope-*` commands
until they had visited a cscope buffer.

To address this, a new function `cscope--init` is introduced. This
function encapsulates the logic for generating the search functions,
toggle functions, and transient menu actions.  `cscope--init` is called
both within `cscope-mode` and when the `cscope.el` file is loaded.

This change improves the user experience by making the cscope commands
immediately available after the package is loaded.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>