Limit syntax highlighting and match highlight for large result sets
This commit introduces a new customization variable,
`cscope-highlight-and-font-line-limit`, to enhance performance when
handling large sets of search results. When the number of lines exceeds
this limit, both syntax highlighting (`cscope-fontify-code-line`) and match
highlighting (`cscope-highlight-match`) are disabled. This change addresses
potential slowdowns that occur due to extensive syntax processing in large
result sets.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>