Add functions to manage error buffers in cscope mode
This commit introduces two new functions, `cscope-quit-current-error` and
`cscope-kill-current-error-buffer`, designed to enhance error buffer
management within cscope mode. The `cscope-current-error-buffer` helper
function is implemented to locate the buffer associated with the current
error. The `cscope-quit-current-error` function allows users to close the
window displaying the current error if it is visible, while
`cscope-kill-current-error-buffer` provides a means to directly kill the
buffer of the current error.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>