Add `cscope-quit-buffer` command and entry to transient menu
This commit introduces `cscope-quit-buffer`, a new command to close the
current cscope buffer.  It also adds an entry to the `cscope-entry`
transient menu to conveniently access this command.

The `cscope-quit-buffer` command allows users to quickly dismiss cscope
result windows, improving workflow efficiency.

The implementation uses `cscope-find-buffer` to locate the cscope buffer
associated with the current directory and then uses `quit-window` to
close the window displaying the buffer.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>