Introduce cscope-file-patterns to generate database command
This commit addresses an issue where the cscope database generation
command was hardcoded to search for '*.[chsS]' files. This prevented
users from easily including other file types, such as Makefiles or
Kconfig files, in the cscope database.

The fix introduces a new customizable variable, `cscope-file-patterns`,
which allows users to specify a list of file patterns to be included in
the database. The `cscope-generate-database-command` function now uses
this variable to generate the appropriate `find` command.

Additionally, the README.org file was updated to reflect this change and
provide an example of how to use the `cscope-file-patterns` variable.

The commit also modifies the advice on
`cscope-generate-database-command` to pass an argument when
`cscope-file-patterns` is customized to prevent searching for 'cscope'
in Makefiles when not appropriate.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>