Prevent concurrent cscope searches
This commit introduces a lock mechanism to prevent concurrent cscope
searches. Previously, multiple cscope queries could be initiated
simultaneously, leading to unpredictable behavior. A `cscope-lock'
variable is introduced. Before initiating a cscope query, the code now
checks if `cscope-lock' is non-nil. If it is, a message is displayed,
and the query is aborted. `cscope-lock' is set to t before sending the
query to the cscope process and set to nil when the cscope process
finishes, as indicated by the "^>>" marker in the output.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>