Robustify `cscope-re-render-context' to handle special characters
The `cscope-re-render-context' function was not correctly re-rendering
the code context when the file path contained special characters,
because the regexp was not robust enough.

This commit introduces `cscope-match-loc' to extract the file and line
number information from the compilation error structure, and uses this
information to build a more robust regular expression. This ensures that
the function correctly identifies and re-renders the code context lines,
even when the file path contains special characters.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>