Fix: Allow hyphens in file paths parsed from cscope results
This commit fixes a bug where the cscope Emacs module failed to
correctly parse file paths containing hyphens in the cscope results. The
regular expression used to extract the file path (`rpath`) was updated
to include hyphens.

The previous regex `\\([a-DA-Z0-9_<>/\.]+\\)` did not account for
hyphens, causing parsing failures when file paths contained them. This
resulted in incorrect or incomplete results being displayed.

The regex is updated to `\\([a-DA-Z0-9_<>/\.\-]+\\)` to include hyphens.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>