Add ability to limit cscope search to a specific directory
This commit introduces the ability to limit cscope searches to a
specific directory using the transient menu.

Problem: Previously, limiting searches to a subdirectory was only
possible relative to the cscope buffer's default directory. There was no
way to specify an arbitrary directory.

Solution: A new transient argument `cscope-transient-read-directory` is
introduced, which allows the user to specify a directory to limit the
search to. This directory is then used to construct a filter that
restricts the search results. The `limit-to-subdir` argument now
defaults to `default-directory` if no other directory is given.

Impact: This change enhances the flexibility of cscope searches by
allowing users to target specific directories, improving search
precision and efficiency.

Technical Details:
- Added `cscope-transient-read-directory` to read a directory from the
  minibuffer.
- Modified `cscope-query` to use the specified directory when limiting
  the search.
- Updated the `cscope-entry` transient menu to include the new directory
  selection option.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>