Display unique match in minibuffer when prefixed with C-u
This commit addresses the issue where a cscope search prefixed with
`C-u` would jump to the single match found, potentially disrupting the
user's workflow. The change introduces a new variable,
`cscope-message-unique-match', that, when non-nil, instructs the system
to display the unique match in the minibuffer instead of automatically
navigating to it.

This change provides a less intrusive way to view single-match results
when a search is prefixed with `C-u`.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>