Correct character class in parsing regular expression
The parsing regular expression used in `cscope-filter` had an incorrect
character class that prevented it from correctly parsing cscope output
containing lowercase characters in file paths and function names.

This commit fixes this issue by updating the character class to include
lowercase characters.  This ensures that cscope results are correctly
parsed, regardless of the case of characters in the file paths or
function names.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>