Prevent grep-regexp-alist corruption by using copy-tree
This commit fixes an issue in the `cscope-mode` initialization related to
the copying of `grep-regexp-alist`. Previously, `copy-list` was used, which
only performs a shallow copy, potentially leading to unintended side
effects when modifying the copied list. By switching to `copy-tree`, a deep
copy is ensured, preventing changes in `cscope-mode` from affecting the
original `grep-regexp-alist`.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>