Add cscope call tree call trace to kill ring
This commit introduces a new feature to copy the call trace of a node in
the cscope call tree to the kill ring. This is useful for documenting or
debugging code by easily capturing the sequence of function calls
leading to a specific point.

This change enhances the cscope call tree navigation by providing a
convenient way to capture and utilize the call trace information.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>